# backend/api/api.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
...
paths:
...
  /api/v1/verify-book-request/accept:
    post:
      tags:
        - VerifyBookRequest
      operationId: "AcceptVerifyBookRequest"
      summary: Accept verify book request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptBookChapterRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
components:
  schemas:
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
    BadRequestResponse:
      type: object
      properties:
        message:
          type: string
    UnauthorizedResponse:
      type: object
      properties:
        message:
          type: string
    NotFoundResponse:
      type: object
      properties:
        message:
          type: string
...
    AcceptBookChapterRequest:
      type: object
      properties:
        verifyBookRequestId:
          type: string
          format: uuid
        accept:
          type: boolean
      required:
        - verifyBookRequestId
        - accept
...